refactor: ManageableAxiosError 타입 분리
